rename log to customlog
