+++ customlog/log_ext.go
@@ -1,6 +1,6 @@
 // (c) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
-package log
+package customlog
 
 import ethtypes "github.com/ava-labs/libevm/core/types"
