+++ eth/filters/filter.go
@@ -33,7 +33,7 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/coreth/core/bloombits"
-	customlog "github.com/ava-labs/coreth/log"
+	"github.com/ava-labs/coreth/customlog"
 	"github.com/ava-labs/coreth/rpc"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/core/types"
@@ -377,7 +377,7 @@ func includes[T comparable](things []T, element T) bool {
 
 // filterLogs creates a slice of logs matching the given criteria.
 func filterLogs(logs []*types.Log, fromBlock, toBlock *big.Int, addresses []common.Address, topics [][]common.Hash) []*types.Log {
-	var check = func(log *types.Log) bool {
+	check := func(log *types.Log) bool {
 		if fromBlock != nil && fromBlock.Int64() >= 0 && fromBlock.Uint64() > log.BlockNumber {
 			return false
 		}
